# Conflicts:
#	hermes-websites-backend-service.iml
#	src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/ImageControllerImplementation.java
#	src/main/java/ro/societateahermes/backendservice/entities/Event.java
#	src/main/java/ro/societateahermes/backendservice/entities/ImageType.java
#	src/main/java/ro/societateahermes/backendservice/entities/dto/CDMemberDTO.java
#	src/main/java/ro/societateahermes/backendservice/service/ImageServiceInterface.java

+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/CDMemberControllerImplementation.java
@@ -1,22 +1,24 @@
 package ro.societateahermes.backendservice.controller.controllerImplementation;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
-
 import ro.societateahermes.backendservice.controller.CDMemberControllerInterface;
-import ro.societateahermes.backendservice.utils.mapper.CDMemberMapper;
 import ro.societateahermes.backendservice.entities.dto.CDMemberDTO;
+import ro.societateahermes.backendservice.exceptions.UnathorizeException;
+import ro.societateahermes.backendservice.service.CDMemberServiceInterface;
 import ro.societateahermes.backendservice.service.serviceImplementation.CDMemberServiceImplementation;
+import ro.societateahermes.backendservice.utils.PermissionChecker;
+import ro.societateahermes.backendservice.utils.RolesActiveUser;
 
 import javax.validation.Valid;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/cd-member")
 public class CDMemberControllerImplementation implements CDMemberControllerInterface {
-    private final CDMemberServiceImplementation cdMemberService;
+    private final CDMemberServiceInterface cdMemberService;
 
     public CDMemberControllerImplementation(CDMemberServiceImplementation cdMemberService) {
         this.cdMemberService = cdMemberService;
@@ -24,20 +26,32 @@ public class CDMemberControllerImplementation implements CDMemberControllerInter
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void saveCDMember(@Valid @RequestBody CDMemberDTO cdMemberDTO) {
-        cdMemberService.save(CDMemberMapper.cdMemberDTOToCDMember(cdMemberDTO));
+    public ResponseEntity<String> saveCDMember(@Valid @RequestBody CDMemberDTO cdMemberDTO) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.checkAdmin(roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
+        if (cdMemberService.isValid(cdMemberDTO)) {
+            cdMemberService.save(cdMemberDTO);
+            return new ResponseEntity<>("", HttpStatus.CREATED);
+        } else {
+            return new ResponseEntity<>("Invalid cd member.", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping
     @Override
     public List<CDMemberDTO> getAllCDMembers() {
-        return cdMemberService.getAllCDMembers().stream()
-                .map(CDMemberMapper::cdMemberToCDMemberDTO).collect(Collectors.toList());
+        return cdMemberService.getAllCDMembers();
     }
 
     @DeleteMapping("/{cd-id}")
     @Override
-    public void deleteCDMember(@PathVariable("cd-id") Long cdMemberID) {
+    public void deleteCDMember(@PathVariable("cd-id") Long cdMemberID) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.checkAdmin(roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         cdMemberService.delete(cdMemberID);
     }
 
@@ -45,7 +59,11 @@ public class CDMemberControllerImplementation implements CDMemberControllerInter
     @ResponseStatus(HttpStatus.OK)
     @Transactional
     @Override
-    public void updateCDMember(@Valid @RequestBody CDMemberDTO cdMemberDTO) {
-        cdMemberService.update(CDMemberMapper.cdMemberDTOToCDMember(cdMemberDTO));
+    public void updateCDMember(@Valid @RequestBody CDMemberDTO cdMemberDTO) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.checkAdmin(roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
+        cdMemberService.update(cdMemberDTO);
     }
 }
